% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinmetrics.R
\name{coinmetrics_asset_metrics}
\alias{coinmetrics_asset_metrics}
\title{Returns requested metrics for specified assets using Coin Metrics V4 API}
\usage{
coinmetrics_asset_metrics(token, metric, start_time, end_time)
}
\arguments{
\item{token}{A stablecoin token}

\item{start_time}{Start of the time interval}

\item{end_time}{End of the time interval}
}
\value{
a dataframe with the results of the query
}
\description{
Returns requested metrics for specified assets using Coin Metrics V4 API
}
\examples{
\donttest{
coinmetrics_asset_metrics(token = 'USDC', metric = 'SplyAct1d', start_time = '2020-01-01', end_time = '2021-07-14')
}
}
