% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current-supply.R
\name{current_supply}
\alias{current_supply}
\alias{current_supply_all}
\title{Retrieve current supply balances on supported blockchains}
\usage{
current_supply(network, token)

current_supply_all(selected_tokens = NULL)
}
\arguments{
\item{network}{the blockchain network}

\item{token}{the stablecoin token}

\item{selected_tokens}{a single or multiple tokens}
}
\description{
\code{current_supply} retrieves the current circulating supply of a stablecoin on
a single blockchain

\code{current_supply_all} retrieves the current circulating supply of a stablecoin
or multiple stablecoins on all underlying blockchains
}
\examples{
\donttest{
current_supply(network = "Ethereum", token = "USDC")
current_supply_all()
current_supply_all("UDSC")
current_supply_all(selected_tokens = c("USDC", "USDT"))
}
}
