% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usdc-current.R
\name{fetch_supply_tron}
\alias{fetch_supply_tron}
\title{Fetch the current USDC circulation on TRON}
\usage{
fetch_supply_tron()
}
\value{
the current circulating amount of USDC
}
\description{
Queries the \code{tronscan} explorer to retrieve the TRC-20 balance of USDC
}
\examples{
fetch_supply_tron()
}
