% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical-supply.R
\name{historical_supply}
\alias{historical_supply}
\alias{historical_supply_all}
\title{Retrieve historical stablecoin circulation}
\usage{
historical_supply(network, token, start_time, end_time)

historical_supply_all(start_time, end_time)
}
\arguments{
\item{network}{The blockchain network}

\item{token}{The stablecoin token}

\item{start_time}{Start of the time interval}

\item{end_time}{End of the time interval}
}
\value{
a dataframe with the historical supply
}
\description{
Queries the Coin Metrics \code{asset-metrics} endpoint to retrieve historical stablecoin circulation
}
\examples{
\donttest{
historical_supply(network = 'Ethereum', token = 'USDC', start_time = '2020-01-01', end_time = '2021-07-14')
historical_supply_all(start_time = '2020-01-01', end_time = '2021-07-14')
}
}
