% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_current_supply}
\alias{plot_current_supply}
\title{Plot the current stablecoin supply on supported blockchains}
\usage{
plot_current_supply()
}
\value{
a ggplot2 chart showing the current circulating amount of USDC
}
\description{
Retrieves the supply for \code{selected_tokens} on each supported blockchain and
then plots the result on a bar chart
}
\examples{
\donttest{
usdc_plot_current_supply()
}
}
