% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plot_historical_supply}
\alias{plot_historical_supply}
\title{Plot historical stablecoin supply on Ethereum}
\usage{
plot_historical_supply(start_time, end_time)
}
\arguments{
\item{start_time}{Start of the time interval}

\item{end_time}{End of the time interval}
}
\value{
a ggplot2 chart showing the historical circulating amount of USDC on Ethereum
}
\description{
Retrieves the supply for \code{selected_tokens} on each supported blockchain and
then plots the result on a bar chart
}
\examples{
\donttest{
usdc_plot_historical_supply(start_time = '2020-01-01', end_time = '2021-07-18')
}
}
